backend/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.111.0
uvicorn[standard]==0.30.1 # [standard] includes httptools and websockets for better performance

# Database ORM and PostgreSQL driver
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Database migration tool
alembic==1.13.1 # Or the latest stable version you prefer, e.g., 1.16.2 as of my last check

# Environment variable management
python-dotenv==1.0.1
pydantic-settings==2.3.3 # For loading settings from environment variables

# Password hashing
passlib[bcrypt]==1.7.4

# JWT authentication
python-jose==3.3.0

# WebSocket library
websockets==12.0 # For the Contrails listener

# Asynchronous rate limiting
asyncio-throttle==1.0 # Optional, for rate limiting if needed

# --- Optional Additions (Uncomment if needed) ---
asyncpg==0.29.0 # If you plan to use SQLAlchemy's async capabilities with PostgreSQL
httpx==0.27.0   # For making asynchronous HTTP requests to external services
# email-validator==2.1.1 # For robust email field validation

# Image generation
Pillow==10.4.0

# Image generation
Pillow==10.4.0
